fix: Tax Withholding Category Description default
This commit fixes a bug that happens when a Purchase Invoice uses a
Tax Withholding Category without a category_name. The category_name is
used as the tax description which is a required field. The bug was
fixed by using the Tax Withholding Category's name as the description
if the category_name is empty.